+++ network.h
@@ -19,16 +19,16 @@ class Network{
     bool dropout;
     void construct_layers(LayerDescriptor **desc);
     inline void backpropagate(MNIST_data *training_data, Layers_features **nabla);
-    void update_weights_and_biasses(MNIST_data **training_data, int training_data_len, int total_trainingdata_len, double learning_rate, double regularization_rate);
+    void update_weights_and_biasses(MNIST_data **training_data, int training_data_len, int total_trainingdata_len, float learning_rate, float regularization_rate);
     inline void remove_some_neurons(MatrixData ***w_bckup, MatrixData ***b_bckup, int **layers_bckup, int ***indexes);
     inline void add_back_removed_neurons(MatrixData **w_bckup, MatrixData **b_bckup, int *layers_bckup, int **indexes);
     inline void feedforward(MatrixData **input);
-    double cost(MatrixData &required_output, int req_outp_indx);
+    float cost(MatrixData &required_output, int req_outp_indx);
     public:
     OpenclSetup openclenv;
     void store(char *filename);
-    void stochastic_gradient_descent(MNIST_data **training_data, int epochs, int minibatch_len, double learning_rate, bool monitor_learning_cost = false,
-                                    double regularization_rate = 0, MNIST_data **test_data = NULL, int minibatch_count = 500, int test_data_len = 10000,  int trainingdata_len = 50000);
+    void stochastic_gradient_descent(MNIST_data **training_data, int epochs, int minibatch_len, float learning_rate, bool monitor_learning_cost = false,
+                                    float regularization_rate = 0, MNIST_data **test_data = NULL, int minibatch_count = 500, int test_data_len = 10000,  int trainingdata_len = 50000);
     Network(int layers_num, LayerDescriptor **layerdesc, int input_row, int input_col = 1, int input_channel_count = 1,
             int costfunction_type = CROSS_ENTROPY_CF, bool dropout = false);
     Network(char *data);
